food_delivery_backend/requirements.txt
Django==3.2.7
djangorestframework==3.12.4
django-cors-headers==3.8.0
pillow==10.2.0
django-filter==2.4.0
django-extensions==3.2.3
mysqlclient==2.2.4
Faker==23.2.1
djangorestframework-simplejwt==5.0.0
phonenumbers==8.12.31
channels==3.0.4
